Refactor: Separate dbService into multiple repositories/isolated-services for improved maintainability
