+++ packages/commons/src/repositories/db.ts
@@ -4,6 +4,7 @@ import pgPromise, {
   IColumnDescriptor,
   IDatabase,
   IMain,
+  ITask,
 } from "pg-promise";
 import {
   IClient,
@@ -11,7 +12,7 @@ import {
 } from "pg-promise/typescript/pg-subset.js";
 
 export type DB = IDatabase<unknown>;
-export type { IMain, IClient, ColumnSet, IColumnDescriptor };
+export type { IMain, IClient, ColumnSet, IColumnDescriptor, ITask };
 
 export function initDB({
   username,
