+++ packages/jwt-audit-analytics-writer/src/model/domain/errors.ts
@@ -1,34 +1,16 @@
 import { InternalError } from "pagopa-interop-kpi-models";
 
 export const errorCodes = {
-  insertStagingRecordsError: "INSERT_STAGING_RECORDS_ERROR",
-  mergeDataError: "MERGE_DATA_ERROR",
-  deduplicateStagingRecordsError: "DEDUPLICATE_STAGING_RECORDS_ERROR",
+  setupStagingTablesError: "SETUP_STAGING_TABLES_ERROR",
 } as const;
 
 export type ErrorCodes = keyof typeof errorCodes;
 
-export function insertStagingRecordsError(
+export function setupStagingTablesError(
   detail: unknown
 ): InternalError<ErrorCodes> {
   return new InternalError({
-    detail: `Database error inserting staging records: ${detail}`,
-    code: "insertStagingRecordsError",
-  });
-}
-
-export function mergeDataError(detail: unknown): InternalError<ErrorCodes> {
-  return new InternalError({
-    detail: `Database error merging data: ${detail}`,
-    code: "mergeDataError",
-  });
-}
-
-export function deduplicateStagingRecordsError(
-  detail: unknown
-): InternalError<ErrorCodes> {
-  return new InternalError({
-    detail: `Database error deduplicating staging records data: ${detail}`,
-    code: "deduplicateStagingRecordsError",
+    detail: `Database error occurred while setting up staging tables. ${detail}`,
+    code: "setupStagingTablesError",
   });
 }
